finally after testing with my pears everything is correct

+++ Cpp09/ex00/inc/btc.hpp
@@ -6,7 +6,7 @@
 /*   By: eseferi <eseferi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/25 12:02:39 by eseferi           #+#    #+#             */
-/*   Updated: 2024/03/25 17:06:24 by eseferi          ###   ########.fr       */
+/*   Updated: 2024/03/25 17:51:52 by eseferi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 #include <exception>
 #include <algorithm>
 #include <string>
-#include <sstream>
 #include <ctime>
 #include <cstring>
 #include <map>
@@ -25,5 +24,5 @@ std::string removeWhitespace(const std::string& str);
 bool	checkDate(struct tm& tm, const std::string& date);
 
 /* parsing.cpp */
-bool parseDbCsv(std::ifstream& ifs_csv, std::map<std::string, float>& dbCsv);
-bool parse(char **argv, std::map<std::string, float>& dbCsv);
+bool parseDbCsv(std::ifstream& ifs_csv, std::map<std::string, double>& dbCsv);
+bool parse(char **argv, std::map<std::string, double>& dbCsv);
